Updated Jet Lancer to arch64
