--- ports/jetlancer/jetlancer/config.ini.default
@@ -1,57 +1,0 @@
-[save]
-last_played_save="1.000000"
-[firstrun]
-firstrun="0.000000"
-[game]
-rank_display="1.000000"
-language="0.000000"
-camera_shake="0.000000"
-text_speed="0.000000"
-disable_loadout_camera_rotation="0.000000"
-score_popups_enabled="0.000000"
-[custom]
-hud_red="255.000000"
-hud_green="255.000000"
-hud_blue="255.000000"
-hud_alpha="0.600000"
-skip_intro="0.000000"
-[video]
-vsync="0.000000"
-fullscreen="1.000000"
-effects="0.000000"
-hdscaling="0.000000"
-shaders="0.000000"
-[audio]
-mas_volume="1.000000"
-sfx_volume="0.500000"
-bgm_volume="0.250000"
-[keyboard]
-key_type="0.000000"
-key_banktype="1.000000"
-key_trackball="0.000000"
-key_bankleft="65.000000"
-key_bankright="68.000000"
-key_thrust="87.000000"
-key_aftrbrn="32.000000"
-key_roll="37.000000"
-key_minigun="38.000000"
-key_weapon="39.000000"
-key_weaponalt="69.000000"
-key_weapon0="49.000000"
-key_weapon1="50.000000"
-key_weapon2="51.000000"
-key_weapon3="52.000000"
-[gamepad]
-pad_invert="0.000000"
-pad_vibration="1.000000"
-pad_banktype="1.000000"
-pad_deadzone="0.300000"
-pad_thrust="32776.000000"
-pad_aftrbrn="32775.000000"
-pad_roll="32771.000000"
-pad_minigun="32769.000000"
-pad_weapon="32773.000000"
-pad_weapon0="32781.000000"
-pad_weapon1="32784.000000"
-pad_weapon2="32782.000000"
-pad_weapon3="32783.000000"
