+++ ports/jetlancer/jetlancer/gmloader.json
@@ -1,0 +1,9 @@
+{
+    "save_dir" : "saves",
+    "apk_path" : "game.port",
+    "show_cursor" : false,
+    "disable_controller" : false,
+    "disable_depth" : false,
+    "disable_rumble" : false,
+    "force_platform" : "os_windows"
+}
