+++ ports/jetlancer/port.json
@@ -1,5 +1,5 @@
 {
-  "version": 3,
+  "version": 4,
   "name": "jetlancer.zip",
   "items": [
     "Jet Lancer.sh",
@@ -12,20 +12,36 @@
       "Fraxinus88",
       "kotzebuedog"
     ],
-    "desc": "Jet Lancer is a high-velocity aerial combat shooter where you duel enemy aviators, hunt giant robots, and defy death at supersonic speeds as you evolve from an unknown rookie aviator into a once-in-a-generation ACE pilot.",
-    "desc_md": "Jet Lancer is a high-velocity aerial combat shooter where you duel enemy aviators, hunt giant robots, and defy death at supersonic speeds as you evolve from an unknown rookie aviator into a once-in-a-generation ACE pilot.\n\n**Note: This is version 1.0.17!**",
-    "inst": "Purchase the game from steam or gog.\nHead to https://portmaster.games/ on the specific game files to place in ports/jetlancer/gamedata",
-    "inst_md": "Purchase the game from [Steam](https://store.steampowered.com/app/913060/Jet_Lancer/) OR [GOG](https://www.gog.com/en/game/jet_lancer)\n\n**Steam**: Download the latest version of the game or use steam console and type download_depot 913060 913061 5560224757383287130, copy over all the game file to ports/jetlancer/gamedata\n\n**GOG**: Download Version 1.0.17. Do this by going to More > Manage Installation > Configure > Uncheck auto-updates > Show more versions > Select Version 1.0.17. Copy over all the game file to ports/jetlancer/gamedata",
+    "desc": "Jet Lancer is a high-velocity aerial combat shooter where you duel enemy aviators, hunt giant robots, and defy death at supersonic speeds as you evolve from an unknown rookie aviator into a once-in-a-generation ACE pilot. ArkOS 1gb devices may crash because of running out of ram!",
+    "desc_md": "Jet Lancer is a high-velocity aerial combat shooter where you duel enemy aviators, hunt giant robots, and defy death at supersonic speeds as you evolve from an unknown rookie aviator into a once-in-a-generation ACE pilot.\n\n**Note: This is version 1.0.17! ArkOS 1gb devices may crash because of running out of ram!**",
+    "inst": "Purchase the game from steam or gog.\nHead to https://portmaster.games/ on the specific game files to place in ports/jetlancer/assets",
+    "inst_md": "Purchase the game from [Steam](https://store.steampowered.com/app/913060/Jet_Lancer/) OR [GOG](https://www.gog.com/en/game/jet_lancer). \n**Steam**: Download the latest version of the game or use steam console and type download_depot 913060 913061 5560224757383287130, copy over all the game file to ports/jetlancer/assets \n**GOG**: Download Version 1.0.17. Do this by going to More > Manage Installation > Configure > Uncheck auto-updates > Show more versions > Select Version 1.0.17. Copy over all the game file to ports/jetlancer/assets",
     "genres": [
       "action"
     ],
     "image": null,
     "rtr": false,
     "exp": false,
-    "runtime": null,
-    "reqs": [],
+    "runtime": [],
+    "store": [
+      {
+        "name": "Steam",
+        "gameurl": "https://store.steampowered.com/app/913060/Jet_Lancer/",
+        "developerurl": ""
+      },
+      {
+        "name": "GOG",
+        "gameurl": "https://www.gog.com/en/game/jet_lancer",
+        "developerurl": ""
+      }
+    ],
+    "availability": "paid",
+    "reqs": [
+      "power"
+    ],
     "arch": [
-      "armhf"
-    ]
+      "aarch64"
+    ],
+    "min_glibc": ""
   }
 }
